get attractions and foods for admin

+++ app/Http/Controllers/Auth/AuthController.php
@@ -47,7 +47,7 @@ class AuthController extends Controller
 
     public function logout(Request $request) {
         $user = $request->user();
-        $user->currentAccessToken()->delete();
+        if($user) $user->currentAccessToken()->delete();
 
         return response('', 204);
     }
